Migrate default values, catch migration errors
It is possible that data migration fails, e.g. because the
Oracle table has a data type that is not supported by oracle_fdw.

Then all subsequent actions on that table, like primary key creation,
are bound to fail as well.

Since the migration should proceed anyway, such errors must be caught
and reported.