Refractor code to introduce "oracle_migrate_refresh"
When working with an Oracle database that is a "moving target",
it can be helpful to refresh the values in the PostgreSQL stage with the
current values from the Oracle stage.

This is easily done by separating the definitions of the tables and the
INSERT statements into different functions and turning INSERT into
INSERT ... ON CONFLICT.  The definition remains in "oracle_migrate_prepare",
which calls the new function "oracle_migrate_refresh" that polulates the tables.