Rewrite ora_migrator to be a db_migrator plugin
A lot of the functionality is moved to db_migrator, only the
Oracle-specific functionality remains in the plugin.

Along the way, several bugs and inconsistencies were found and fixed.
(Hopefully not too many bugs were introduced.)

The "oracle_code_count" function is gone, and the view
"migration_cost_estimate" has changed considerably.